requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
jinja2==3.1.2
aiosmtplib==2.0.2
click==8.0.0
pydantic-settings==2.0.0
python-json-logger==2.0.7
httpx==0.24.1
supabase==1.2.0
asyncpg==0.28.0
